None

+++ delete.py
@@ -15,7 +15,7 @@ while True:
     if want == "delete unimportant files":
         files = glob.glob(os.getcwd() + "/*.*")
         for f in files:
-            if not f.endswith('users.json') and not f.endswith('delete.py') and not f.endswith('commands.py') and not f.endswith('config.py') and not f.endswith('main.py'):
+            if not f.endswith('users.json') and not f.endswith('delete.py') and not f.endswith('commands.py') and not f.endswith('config.py') and not f.endswith('main.py') and not f.endswith('db.sql'):
                 os.remove(f)
         print("All files was deleted\n")
         find = 1
